requirements.txt
Flask
pandas
snowflake-connector-python
